Added decimal numbers on the DownloadPage and DownloadItem views to report download progress and speed.

+++ src/PCGamingModApp.Core/Helpers/ConversionHelper.cs
@@ -4,16 +4,16 @@ namespace PCGamingModApp.Core.Helpers;
 
 public static class ConversionHelper
 {
-    public static long ConvertBytesToUnit(long bytes, SizeUnit unit)
+    public static double ConvertBytesToUnit(long bytes, SizeUnit unit)
     {
         return unit switch
         {
-            SizeUnit.KB => bytes / 1024L,
-            SizeUnit.MB => bytes / (1024L * 1024L),
-            SizeUnit.GB => bytes / (1024L * 1024L * 1024L),
-            SizeUnit.TB => bytes / (1024L * 1024L * 1024L * 1024L),
-            SizeUnit.PB => bytes / (1024L * 1024L * 1024L * 1024L * 1024L),
-            SizeUnit.EB => bytes / (1024L * 1024L * 1024L * 1024L * 1024L * 1024L),
+            SizeUnit.KB => bytes / 1024d,
+            SizeUnit.MB => bytes / (1024d * 1024),
+            SizeUnit.GB => bytes / (1024d * 1024 * 1024),
+            SizeUnit.TB => bytes / (1024d * 1024 * 1024 * 1024),
+            SizeUnit.PB => bytes / (1024d * 1024 * 1024 * 1024d * 1024),
+            SizeUnit.EB => bytes / (1024d * 1024 * 1024 * 1024 * 1024 * 1024),
             _ => bytes
         };
     }
